+++ src/Paprika.Benchmarks/FixedMapBenchmarks.cs
@@ -1,6 +1,7 @@
 ﻿using System.Buffers.Binary;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Disassemblers;
+using Paprika.Db;
 using Paprika.Pages;
 
 namespace Paprika.Benchmarks;
