+++ src/Paprika.Tests/DataPageTests.cs
@@ -3,6 +3,7 @@ using FluentAssertions;
 using Nethermind.Int256;
 using NUnit.Framework;
 using Paprika.Crypto;
+using Paprika.Db;
 using Paprika.Pages;
 using static Paprika.Tests.Values;
