+++ src/Paprika.Tests/NibblePathTests.PrefixTests.cs
@@ -1,4 +1,5 @@
 ﻿using NUnit.Framework;
+using Paprika.Db;
 
 namespace Paprika.Tests;
