+++ src/Paprika.Tests/PrinterTests.cs
@@ -1,6 +1,7 @@
 ﻿using Nethermind.Int256;
 using NUnit.Framework;
 using Paprika.Db;
+using Paprika.Development;
 using Paprika.Pages;
 
 using static Paprika.Tests.Values;
