--- src/Paprika/CommitOptions.cs
@@ -1,19 +1,0 @@
-﻿namespace Paprika;
-
-public enum CommitOptions
-{
-    /// <summary>
-    /// Flushes db only once, ensuring that the data are stored properly.
-    /// The root is stored ephemerally, waiting for the next commit to be truly stored.
-    ///
-    /// This guarantees ATOMIC (from ACID) but not DURABLE.  
-    /// </summary>
-    FlushDataOnly,
-
-    /// <summary>
-    /// Flush data, then the root.
-    ///
-    /// This guarantees ATOMIC and DURABLE (from ACID).
-    /// </summary>
-    FlushDataAndRoot
-}
