+++ src/Paprika/Pages/DataPage.cs
@@ -1,7 +1,7 @@
-﻿using System.Buffers.Binary;
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using Paprika.Db;
 
 namespace Paprika.Pages;
