+++ src/Paprika/Pages/FanOut256Page.cs
@@ -1,6 +1,7 @@
 ﻿using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using Paprika.Db;
 
 namespace Paprika.Pages;
