+++ src/Paprika/Pages/FixedMap.cs
@@ -4,6 +4,8 @@ using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using Paprika.Crypto;
+using Paprika.Db;
+using Paprika.Development;
 
 namespace Paprika.Pages;
