--- src/Paprika/Pages/PageExtensions.cs
@@ -1,18 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-namespace Paprika.Pages;
-
-public static class PageExtensions
-{
-    public static void CopyTo<TPage>(this TPage page, TPage destination) where TPage : unmanaged, IPage =>
-        Unsafe.As<TPage, Page>(ref page).Span.CopyTo(Unsafe.As<TPage, Page>(ref destination).Span);
-
-    public static void CopyTo<TPage>(this TPage page, Page destination) where TPage : unmanaged, IPage =>
-        Unsafe.As<TPage, Page>(ref page).Span.CopyTo(destination.Span);
-
-    public static Page AsPage<TPage>(this TPage page) where TPage : unmanaged, IPage =>
-        Unsafe.As<TPage, Page>(ref page);
-
-    public static TPage Cast<TPage>(this Page page) where TPage : unmanaged, IPage =>
-        Unsafe.As<Page, TPage>(ref page);
-}
